refactor: restructure profile editing components and update form handling
- Removed the old ProfileEditMain component and replaced it with a new ProfileEditMainForm for better organization and functionality.
- Updated ProfileSidebarEdit to use a new form structure for managing tech stacks and social links.
- Enhanced the PinnedProjects component to use updated project properties.
- Introduced a new useProfileUpdate hook for handling profile updates with improved error handling and user feedback.
- Adjusted profile schema to reflect changes in social links structure.
